Refactor format of data tables to be in JSON, not Typescript
And relocate them to not be in a nested directory. This should be more
efficient for loading those modules, not having to instantiate a large
number of Decimal objects that won't actually be touched.

Part of the motivation/approach is to rewrite the dynamic imports to
be more conducive to bundlers that can now infer which files might be
dynamically imported by having the filename (inc extension) in the
import statement itself.